Really good UniqueID system done, began work on preplaced objects
The new NetworkIdentity class uses a helper class called UniqueID, which stores both a prefab id and a prefab instance id

+++ unity-project/Assets/LiteNetworking/Entity/EntityPackets.cs
@@ -3,7 +3,7 @@ using LiteNetworking;
 using UnityEngine;
 public class SpawnEntityPacket : LitePacket {
     public int entityId;
-    public int prefabId;
+    public long prefabId;
     public int authority;
     public Vector3 position;
